refactor(batch): split costing page helpers
